requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0
openpyxl==3.1.2
python-dotenv==1.0.1
pytest==8.0.2
black==24.2.0
flake8==7.0.0
isort==5.13.2
mypy==1.8.0 